Change fuzz for CI

+++ Tests/PrimitiveTypesTests/ArithmeticTests/FuzzDivTests.swift
@@ -14,7 +14,7 @@ final class FuzzDivRemSpec: QuickSpec {
             }
 
             it("run fuzz for divRem") {
-                let val1 = UInt128.random(in: 2 ... UInt128.max)
+                let val1 = UInt64.random(in: 2 ... UInt64.max)
                 /*
                  var index = 0
                  while index < 10 {
